requirements.txt
python-opencv
python-dotenv
fastapi
uvicorn
openai
numpy